Top stats added.

+++ src/main/java/Data/ResultMapper.java
@@ -24,7 +24,7 @@ public class ResultMapper {
                 String diciplin = resultset.getString("Disciplin");
                 int competitionID = resultset.getInt("CompetitionID");
                 double resultMeters = resultset.getDouble("ResultMeters");
-                int resultTime = resultset.getInt("ResultTime");
+                double resultTime = resultset.getDouble("ResultTime");
                 boolean training = resultset.getBoolean("Training");
 
                 Member member = null;
@@ -51,7 +51,7 @@ public class ResultMapper {
         return tmpResults;
     }
 
-    public Result addNewResult(Member member, String diciplin, Competition competition, Double resultMeters, int resultTime, boolean training){
+    public Result addNewResult(Member member, String diciplin, Competition competition, Double resultMeters, double resultTime, boolean training){
         int memberId = member.getId();
         int competitionId = competition.getId();
 
@@ -64,7 +64,7 @@ public class ResultMapper {
             statement.setString(2,diciplin);
             statement.setInt(3,competitionId);
             statement.setDouble(4,resultMeters);
-            statement.setInt(5,resultTime);
+            statement.setDouble(5,resultTime);
             statement.setBoolean(6,training);
             statement.executeUpdate();
             ResultSet tableKeys = statement.getGeneratedKeys();
